Add loadWithTimeout function to fix long loading
Add a loadWithTimeout function that loads code from a URL. It is
intended to replace loadstring(game:HttpGet("URL"))() with
loadWithTimeout("URL", timeout?).

The function will fetch and load the code in a separate thread that
gets killed if the timeout is reached. This will prevent the script from
taking a very long time to load if the target URL is temporarily down,
as well as catching any errors that may occur during execution.

Add a getAssetUri() function to get the URI of an asset from the user-
specified ID. This ensures that if the user specifies an invalid ID
(e.g., not a number), the script will gracefully output a warning
message and return an empty image instead of erroring.

Add checks to verify if Lucide icon library is loaded before using it.

Make the booster code load and run asynchronously (using task.spawn) to
decrease loading times.