Add ToggleUIKeybind setting to :CreateWindow()
- This allows the script developer to override the keybind for toggling
the UI. Particularly useful for scripts that use the same keybind as
the default UI toggle.

If ToggleUIKeybind is set, it will be used instead of the keybind
stored in the user's configuration file. If the user then changes the
keybind through the settings menu, the override will be removed and
the new keybind will be used and saved as ususal.

- The override system can be used to override any setting that is
configurable by the user.

- Also set the user-facing toggle for usageAnalytics to false if requests
have already been disabled by the developer.